Move lecHomePg to Pages folder and correct structure
